+++ src/commands/cards/get.rs
@@ -26,11 +26,11 @@ use super::non_empty;
 #[derive(Debug, Args)]
 pub struct GetArgs {
     /// Wekan board ID.
-    #[arg(value_parser = non_empty)]
+    #[arg(long = "board", value_parser = non_empty)]
     pub board_id: String,
 
     /// Wekan list ID.
-    #[arg(value_parser = non_empty)]
+    #[arg(long = "list", value_parser = non_empty)]
     pub list_id: String,
 
     /// Wekan card ID.