+++ src/commands/comments/get.rs
@@ -16,11 +16,11 @@ use super::non_empty;
 #[derive(Debug, Args)]
 pub struct GetArgs {
     /// Wekan board ID.
-    #[arg(value_parser = non_empty)]
+    #[arg(long = "board", value_parser = non_empty)]
     pub board_id: String,
 
     /// Wekan card ID.
-    #[arg(value_parser = non_empty)]
+    #[arg(long = "card", value_parser = non_empty)]
     pub card_id: String,
 
     /// Wekan comment ID.