+++ src/commands/lists/list.rs
@@ -14,7 +14,7 @@ use super::non_empty;
 #[derive(Debug, Args)]
 pub struct ListArgs {
     /// Wekan board ID.
-    #[arg(value_parser = non_empty)]
+    #[arg(long = "board", value_parser = non_empty)]
     pub board_id: String,
 }