+++ src/commands/swimlanes/create.rs
@@ -15,7 +15,7 @@ use super::{finite_number, non_empty, trimmed_non_empty};
 #[derive(Debug, Args)]
 pub struct CreateArgs {
     /// Wekan board ID.
-    #[arg(value_parser = non_empty)]
+    #[arg(long = "board", value_parser = non_empty)]
     pub board_id: String,
 
     /// Swimlane title.