+++ src/commands/swimlanes/get.rs
@@ -16,7 +16,7 @@ use super::non_empty;
 #[derive(Debug, Args)]
 pub struct GetArgs {
     /// Wekan board ID.
-    #[arg(value_parser = non_empty)]
+    #[arg(long = "board", value_parser = non_empty)]
     pub board_id: String,
 
     /// Wekan swimlane ID.