+++ src/commands/swimlanes/update.rs
@@ -18,7 +18,7 @@ use super::{non_empty, trimmed_non_empty};
 #[derive(Debug, Args)]
 pub struct UpdateArgs {
     /// Wekan board ID.
-    #[arg(value_parser = non_empty)]
+    #[arg(long = "board", value_parser = non_empty)]
     pub board_id: String,
 
     /// Wekan swimlane ID.